Lab8PSP/requirements.txt
click==7.1.2
Flask==1.1.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
itsdangerous==1.1.0
Jinja2==2.11.2
logging==0.4.9.6
MarkupSafe==1.1.1
pkg-resources==0.0.0
psycopg2-binary==2.8.6
PyYAML==5.3.1
SQLAlchemy==1.3.20
trafaret==2.1.0
Werkzeug==1.0.1
